+++ CHANGES.md
@@ -16,6 +16,7 @@
 * Fix memory leaks in Load(string, Encoding) and Save(string, Encoding).
 * Documentation updates.
 * Minor fixes.
+	* Fixed a bug related to saving to a nonexistant file.
 
 # 1.1.0
 * Basic serialization and deserialization (Serialization namespace).